Add folders + reorder some classes + split classes into their own file
