+++ Client/Frontend/TabContentScripts/FindInPageHelper.swift
