+++ Client/Frontend/TabContentScripts/LocalRequestHelper.swift
