+++ Client/Frontend/TabContentScripts/NightModeHelper.swift
