+++ Client/Frontend/TabContentScripts/NoImageModeHelper.swift
