+++ Client/Frontend/TabContentScripts/PrintHelper.swift
