+++ Client/Frontend/TabContentScripts/SessionRestoreHelper.swift
