+++ Client/Frontend/TabContentScripts/UserScriptManager.swift
